refactor: extract security dialog components
